requirements.txt
numpy
opencv-python
python-can
